pyproject.toml
[project]
name = "django-drf-keycloak-auth-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.2.7",
    "django-cors-headers==4.9.0",
    "django-drf-keycloak-auth==0.0.2",
    "djangorestframework==3.16.1",
    "drf-spectacular==0.28.0",
    "python-dotenv==1.1.1",
    "python-keycloak==5.8.1",
]
